Code migration: expectation.py -> optimization.py
- Move code and functions that aim to reduce and optimize the measurement cost into a new file (`optimization.py`)
- `expectation.py` is now more focused: Only code for obtaining the Hamiltonian expectation value left in there
